+++ CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.9.2] - 2026-07-07
+
+### Fixed
+
+- Pets without beds now drop collar tags on death.
+- Fixed Axolotl data being lost when bucketed.
+- Fixed client-server desync for active pet collar enchantments.
+
 ## [1.9.1] - 2026-07-06
 
 ### Changed